# Please enter a commit message to explain why this merge is necessary,
# especially if it merges an updated upstream into a topic branch.
#
# Lines starting with '#' will be ignored, and an empty message aborts
# the commit.

+++ test/getcheckedindbs.sh
@@ -4,6 +4,8 @@
 # that are currently checked in
 
 while IFS='' read -r f || [[ -n "${f}" ]]; do
+    if [[ ! $f =~ .*sqlschema.* ]]; then
 	rm -f ${f}
 	git checkout ${f}
+    fi
 done < dbfiles.txt
